# docs/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Inventory service",
    "description": "Service for managing inventory",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "InventoryService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/inventory/compensate": {
      "post": {
        "summary": "Компенсация товара",
        "operationId": "InventoryService_CompensateProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/inventoryserviceCompensateProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/inventoryserviceCompensateProductRequest"
            }
          }
        ],
        "tags": [
          "InventoryService"
        ]
      }
    },
    "/inventory/reserve": {
      "post": {
        "summary": "Резервирование товара",
        "operationId": "InventoryService_ReserveProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/inventoryserviceReserveProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/inventoryserviceReserveProductRequest"
            }
          }
        ],
        "tags": [
          "InventoryService"
        ]
      }
    }
  },
  "definitions": {
    "inventoryserviceCompensateProductRequest": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string",
          "format": "int64"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "Сообщение для запроса компенсации товара"
    },
    "inventoryserviceCompensateProductResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "Сообщение для ответа на компенсацию товара"
    },
    "inventoryserviceReserveProductRequest": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string",
          "format": "int64"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "Сообщение для запроса резервирования товара"
    },
    "inventoryserviceReserveProductResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "Сообщение для ответа на резервирование товара"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}